Add commands for managing guild bridging
guilds status will list the guild names, their ids, and whether or not the
guild is bridged.

guilds bridge <guildID> will bridge the guild with the given ID. If you add the
--entire flag, it will create a portal for each channel on the guild.

guilds unbridge <guildID> will stop bridging the guild with the given ID and
remove all portals for that guild.

Refs #8